backend/ai_court/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.6.2
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
google-generativeai>=0.3.0
pydantic>=2.5.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
pypdf2>=3.0.1
python-magic>=0.4.27
aiofiles>=23.2.1
nest-asyncio>=1.5.8
python-dateutil>=2.8.20
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.18
passlib>=1.7.4
elevenlabs>=0.2.26
gtts>=2.4.0
faster-whisper>=0.10.0
